+++ components/remote-controls/categories/impl/src/main/kotlin/com/flipperdevices/remotecontrols/impl/categories/presentation/data/BackendDeviceCategoriesRepository.kt
@@ -2,7 +2,7 @@ package com.flipperdevices.remotecontrols.impl.categories.presentation.data
 
 import com.flipperdevices.core.di.AppGraph
 import com.flipperdevices.core.ktx.jre.FlipperDispatchers
-import com.flipperdevices.ifrmvp.api.infrared.InfraredBackendApi
+import com.flipperdevices.ifrmvp.api.infrared.FlipperInfraredBackendApi
 import com.flipperdevices.ifrmvp.backend.model.DeviceCategory
 import com.squareup.anvil.annotations.ContributesBinding
 import kotlinx.coroutines.withContext
@@ -10,7 +10,7 @@ import javax.inject.Inject
 
 @ContributesBinding(AppGraph::class, DeviceCategoriesRepository::class)
 class BackendDeviceCategoriesRepository @Inject constructor(
-    private val infraredBackendApi: InfraredBackendApi,
+    private val infraredBackendApi: FlipperInfraredBackendApi,
 ) : DeviceCategoriesRepository {
 
     override suspend fun fetchCategories(): Result<List<DeviceCategory>> = runCatching {