+++ components/remote-controls/setup/api/src/main/kotlin/com/flipperdevices/remotecontrols/api/SaveTempSignalApi.kt
@@ -17,7 +17,6 @@ interface SaveTempSignalApi : InstanceKeeper.Instance {
 
     sealed interface State {
         data object Pending : State
-        data object Error : State
         data class Uploading(val progressInternal: Long, val total: Long) : State {
             val progressPercent: Float = if (total == 0L) 0f else progressInternal / total.toFloat()
         }