+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Attention: don't forget to add the flag for F-Droid before release
 - [Feature] Better user-ux when configuring remote control
 - [Feature] Navigate to previous setup item on remote controls
 - [Feature] Add flipper action dialogs into remote control and move it into bottombar
+- [Feature] Add error display into remote controls screens
 - [Feature] Add new icons for remote-controls
 - [Refactor] Load RemoteControls from flipper, emulating animation
 - [Refactor] Update to Kotlin 2.0